nationality_guesser/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.5
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
